requirements.txt
pdf2docx
pymupdf
pdfplumber
streamlit 
streamlit-quill 
pdfplumber
streamlit_ace
pdfkit
weasyprint
streamlit_pdf_viewer
wheel
PyPDF2
python-docx
frontend
pandas 
matplotlib 
seaborn 
psycopg2-binary
spire.pdf
docx2pdf